% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copynumber_mobem.R
\name{get_copynumber_mobem}
\alias{get_copynumber_mobem}
\title{Takes copy number data download from CMP and discretises into a binary event matrix}
\usage{
get_copynumber_mobem(
  cn_data,
  cn_threshold = c("Gain", "Amplification", "Loss", "Deletion"),
  gene_id_col = "symbol",
  model_id_col = "model_id",
  data_types = "WES",
  data_source = "Sanger",
  annot_rownames = NULL,
  events = NULL,
  filename = NULL
)
}
\arguments{
\item{cn_data}{copy number data from CMP download}

\item{cn_threshold}{Category to use for discretising, one of Gain, Amplification, Loss or Deletion.}

\item{gene_id_col}{The name of the column in copy number data to use as the gene identifier. Default symbol.}

\item{model_id_col}{The name of the column in copy number data to use as the model identifier. Default model_id.}

\item{data_types}{Name of the data type to use e.g. WES.}

\item{data_source}{Name of the data source to use. One of Sanger or Broad.}

\item{annot_rownames}{Optional suffix to add to all rownames e.g. Gain for copy number gain events. Default NULL.}

\item{events}{Optional. Set of events to select for final discrete matrix. Default NULL.}

\item{filename}{Optional. Name of file to save matrix as in tsv format. Default NULL.}
}
\value{
Copy Number Binary Event Mutation matrix
}
\description{
Takes copy number data download from CMP and discretises into a binary event matrix
}
\examples{
driver_cn<-"https://cog.sanger.ac.uk/cmp/download/cnv_summary_latest.csv.gz"
library(data.table)
cn_data<-fread(driver_cn)
sanger_wes_Amp<-get_copynumber_mobem(cn_data,cn_threshold="Amplification",annot_rownames="Amplification")

}
