% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fusion_mobem.R
\name{get_fusion_mobem}
\alias{get_fusion_mobem}
\title{Takes fusion data download from CMP and discretises into a binary event matrix}
\usage{
get_fusion_mobem(
  fusion_data,
  start_symbol = "gene_symbol_5prime",
  end_symbol = "gene_symbol_3prime",
  model_id_col = "model_id",
  annot_rownames = NULL,
  events = NULL,
  filename = NULL
)
}
\arguments{
\item{fusion_data}{from CMP download}

\item{start_symbol}{The name of the column in fusion_data to use as the start of fusion identifier. Default gene_symbol_5prime.}

\item{end_symbol}{The name of the column in fusion_data to use as the end of fusion identifier. Default gene_symbol_3prime.}

\item{model_id_col}{The name of the column in fusion_data to use as the model identifier. Default model_id.}

\item{annot_rownames}{Optional suffix to add to all row names e.g. Fusion to denote fusion events. Default NULL.}

\item{events}{Optional. Set of events to select for final discrete matrix. Default NULL.}

\item{filename}{Optional. Name of file to save matrix as in tsv format. Default NULL.}
}
\value{
Fusion Binary Event Matrix
}
\description{
Takes fusion data download from CMP and discretises into a binary event matrix
}
\examples{
all_fusion<-"https://cog.sanger.ac.uk/cmp/download/fusions_latest.csv.gz"
library(data.table)
fusion_data<-fread(all_fusion)
fusion<-get_fusion_mobem(fusion_data,annot_rownames = "fusion")


}
