% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_expression_input.R
\name{get_gene_expression_input}
\alias{get_gene_expression_input}
\title{Convert CMP expression data into matrix format for discretisation.}
\usage{
get_gene_expression_input(
  expr_data,
  filename = "ExpressionMatrix",
  gene_id_col = "ensembl_gene_id",
  model_id_col = "model_id",
  expr_val = "rsem_tpm",
  source = "Sanger",
  log_transform = TRUE,
  save_RDS = TRUE
)
}
\arguments{
\item{expr_data}{copy number data from CMP download}

\item{filename}{Name of file to save matrix as in RDS format. Only used if saveRDS =TRUE. Default ExpressionMatrix.}

\item{gene_id_col}{The name of the column in copy number data to use as the gene identifier. Default ensembl_gene_id.}

\item{model_id_col}{The name of the column in copy number data to use as the model identifier. Default model_id.}

\item{expr_val}{Value to use to populate matrix. Default rsem_tpm.}

\item{log_transform}{Should data be log transformed. Default. TRUE.}

\item{save_RDS}{Save expression matrix to RDS file. Default TRUE.}

\item{data_source}{Name of the data source to use. One of Sanger or Broad.}
}
\value{
Gene expression matrix
}
\description{
Convert CMP expression data into matrix format for discretisation.
}
\examples{
all_expr<-"https://cog.sanger.ac.uk/cmp/download/rnaseq_latest.csv.gz"
library(data.table)
#not run
#expr_data<-fread(all_expr)
#expr_tpm_mat<-get_gene_expression_input(expr_data,save_RDS=FALSE,expr_val="rsem_tpm",log_transform = FALSE)


}
