% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_mobem.R
\name{get_mutation_mobem}
\alias{get_mutation_mobem}
\title{Takes mutation data download from CMP and discretises into a binary event matrix}
\usage{
get_mutation_mobem(
  mutation_data,
  gene_id_col = "ensembl_gene_id",
  model_id_col = "model_id",
  vaf_thresh = 0.15,
  data_types = "WES",
  data_source = "Sanger",
  annot_rownames = NULL,
  events = NULL,
  filename = NULL
)
}
\arguments{
\item{mutation_data}{mutation data from CMP download}

\item{gene_id_col}{The name of the column in mutation data to use as the gene identifier. Default ensembl_gene_id.}

\item{model_id_col}{The name of the column in mutation data to use as the model identifier. Default model_id.}

\item{vaf_thresh}{Threshold to include mutation. Default 0.15.}

\item{data_types}{Name of the data type to use e.g. WES.}

\item{data_source}{Name of the data source to use. One of Sanger or Broad.}

\item{annot_rownames}{Optional suffix to add to all rownames e.g. mutation. Default NULL.}

\item{events}{Optional set of events (mutations) to include in final matrix.}

\item{filename}{Optional. Name of file to save matrix as in tsv format. Default NULL.}
}
\value{
mutation Binary Event Mutation matrix
}
\description{
Takes mutation data download from CMP and discretises into a binary event matrix
}
\examples{
library(data.table)
mutation_file_path<-"https://cog.sanger.ac.uk/cmp/download/mutations_summary_latest.csv.gz"
mutation_data<-fread(mutation_file_path)
sanger_wes<-get_mutation_mobem(mutation_data,annot_rownames = "mut")
broad_wgs<-get_mutation_mobem(mutation_data,data_source="Broad",data_type="WGS",annot_rownames = "mut")


}
