% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_mobem_fromBroad.R
\name{get_mutation_mobem_fromBroad}
\alias{get_mutation_mobem_fromBroad}
\title{Takes mutation data download from Broads depmap.org and discretises into a binary event matrix}
\usage{
get_mutation_mobem_fromBroad(
  broad_mutation_data,
  gene_id_col = "EnsemblGeneID",
  model_id_col = "ModelID",
  variantCol = "VariantInfo",
  GeneCol = "HugoSymbol",
  vaf_thresh = 0.15,
  vafCol = "AF",
  cancerdriverfile,
  drivermutfile,
  cpgmutfile,
  MutCodeFile,
  events = NULL,
  annot_rownames = NULL
)
}
\arguments{
\item{broad_mutation_data}{mutation data from CMP download}

\item{gene_id_col}{The name of the column in mutation data to use as the gene identifier. Default EnsemblGeneID.}

\item{model_id_col}{The name of the column in mutation data to use as the model identifier. Default ModelID.}

\item{variantCol}{Name of the column that contains the variant annotation e.g missense, frameshift, intronic etc}

\item{GeneCol}{Name of the column that corresponds to identifiers used in the cancer driver data e.g. symbol}

\item{vaf_thresh}{Threshold to include mutation. Default 0.15.}

\item{vafCol}{Name of the column in the mutation data containing the VAF values. Default AF.}

\item{cancerdriverfile}{Name of CMP link to cancer drivers annotation file, e.g names of cancer drivers and type Oncogene, LoF, Ambiguous.}

\item{drivermutfile}{Name of CMP link to cancer driver variants file}

\item{cpgmutfile}{Name of CMP link to cancer predisposition mutation file}

\item{MutCodeFile}{Name of csv file containing mappings between Broad variant annotation classes and the Sanger CMP variants annotation.}

\item{events}{Optional. Set of events to select for final discrete matrix. Default NULL.}

\item{annot_rownames}{Optional suffix to add to all rownames e.g. mutation. Default NULL.}
}
\value{
mutation Binary Event Mutation matrix
}
\description{
Takes mutation data download from Broads depmap.org and discretises into a binary event matrix
}
\examples{
library(rtracklayer)
drivermutfile="https://cog.sanger.ac.uk/cmp/download/driver_mutations_latest.csv"
cancerdriverfile="https://cog.sanger.ac.uk/cmp/download/driver_genes_latest.csv"
cpgmutfile="https://cog.sanger.ac.uk/cmp/download/cancer_predisposition_variants_latest.csv"
geneannotfile="https://cog.sanger.ac.uk/cmp/download/gene_identifiers_latest.csv.gz"
MutCodeFile <- system.file("data", "MutationCodes.RData", package = "GenerateMoBEMs")
#subset of the data download from depmap.org used for example purposes:
omicsfile <- system.file("data", "random_sampleOSM.RData", package = "GenerateMoBEMs")
BroadBinary<-get_mutation_mobem_fromBroad(omicsfile,cancerdriverfile = cancerdriverfile,drivermutfile=drivermutfile,cpgmutfile=cpgmutfile,MutCodeFile=MutCodeFile)
}
