% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteomic_mobem.R
\name{get_proteomic_mobem}
\alias{get_proteomic_mobem}
\title{Takes proteomics data download from CMP and discretises into a binary event matrix}
\usage{
get_proteomic_mobem(
  proteomics_data,
  z_thresh = 2,
  direction = c("higher", "lower"),
  gene_id_col = "symbol",
  model_id_col = "model_id",
  annot_rownames = NULL,
  events = NULL,
  filename = NULL
)
}
\arguments{
\item{proteomics_data}{from CMP download}

\item{z_thresh}{Threshold to use for discretising z-score, default 2.}

\item{direction}{Which direction to consider for discretising. One of higher or lower, z-score > z_thresh or z-score < z_thresh respectively}

\item{gene_id_col}{The name of the column in proteomics_data to use as the gene identifier. Default symbol.}

\item{model_id_col}{The name of the column in proteomics_data to use as the model identifier. Default model_id.}

\item{annot_rownames}{Optional suffix to add to all rownames e.g. High for high expressed proteins. Default NULL.}

\item{events}{Optional. Set of events to select for final discrete matrix. Default NULL.}

\item{filename}{Optional. Name of file to save matrix as in tsv format. Default NULL.}
}
\value{
Proteomic Binary Event Matrix
}
\description{
Takes proteomics data download from CMP and discretises into a binary event matrix
}
\examples{
library(data.table)
all_prot<-"https://cog.sanger.ac.uk/cmp/download/proteomics_latest.csv.gz"
proteomics_data<-fread(all_prot)
#Proteomic binary matrices
prot_high<-get_proteomic_mobem(proteomics_data,z_thresh=2.3,annot_rownames = "High")
prot_low<-get_proteomic_mobem(proteomics_data,z_thresh=c(-2.3),direction="lower",annot_rownames = "Low")

}
